Move diff calculation to background thread.
In it's current state there is alot of coupling of test code. This means
I had to go through each test to ensure it would use the
`ImmediateExecutor` I created to enable synchronously tested code. To be
able to inject this executor I needed to open up constructors, which is
why there is now a few different visible constructors for executors.

Although it may have been more appropriate to save this for a different
pull request, I also cleaned up the test-code to try and keep it more
DRY as appropriate. This involved the use of the `@SetUp` method to
rebuild objects for each test.

ISSUE #286